[test] Clean test file - Skip IPFS issues checks
